requirements.txt
pandas
numpy
requests
sqlalchemy
google-cloud-bigquery
streamlit
plotly